server/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
requests
pydantic
